Remove rate limit link creation in RatingProcessor because the RatingAssembler already handles that.
Add hal forms template for rating politicians when a jwt is present in
the findById request for ratings and also if the user is currently not
rate limited on the politician.

Add Spring hateoas assembler for rating politicians in the context of
PoliticiansRating/RatingsController endpoints that is a bean with a
request scope.

Remove redundant test for RatingService on rate limited situations.

Change PoliticianHttpAdapterTest from having a defined port into a
random port to solve mysterious port already in use exception.