Add Score value object. Score is the class for replacing primitive ratings by using doubles. The Score class has validations and scores/ratings can only be greater than 0 or less than 10
Add AverageRating value object that extends on Score to basically get
all the validation that Score has. Essentially, AverageRating is the
same with Score, the difference is that AverageRating can calculate
ratings for you based on total ratings accumulated and the count needed
for dividing with the total ratings. Also it has a better name on it

Rating enums can handle calculation of ratings